requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langgraph
yfinance
finnhub-python
stockstats
pandas
tavily-python
beautifulsoup4
chromadb
rich
tiktoken
